codex-tools: extract collaboration tool specs
